unknown change check

+++ app/Http/Controllers/ListingController.php
@@ -81,6 +81,12 @@ class ListingController extends Controller
     //
     // Update Listing
     public function update(Request $request, Listing $listing) {
+
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',  // some doubt here
@@ -104,6 +110,12 @@ class ListingController extends Controller
     //
     // Delete Listing
     public function destroy(Listing $listing) {
+        
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing Deleted Successfully');
     }
